stream as CRTP. Has issues.

+++ hostapi/Common.h
@@ -102,7 +102,7 @@ void analogWrite(pin_size_t pinNumber, int value);
 unsigned long millis(void);
 unsigned long micros(void);
 void delay(unsigned long);
-void delayMicroseconds(unsigned int us);
+void delayMicroseconds(unsigned long us);
 unsigned long pulseIn(pin_size_t pin, uint8_t state, unsigned long timeout);
 unsigned long pulseInLong(pin_size_t pin, uint8_t state, unsigned long timeout);
